Combine the removeItem and removeSep methods.
I did this because i felt it might be a little confusing what list item
your removing. This way you specify what list item your removing. It
better matches the adding methods because each of thoes you can specify
the exact location in the menu you are adding it at.